Moved validation utility into common library.
This way, it can be used in other servlets.
