+++ kangaroo-servlet-oauth2/src/main/java/net/krotscheck/kangaroo/servlet/oauth2/resource/grant/RefreshTokenGrantHandler.java
@@ -24,7 +24,7 @@ import net.krotscheck.kangaroo.database.entity.ClientType;
 import net.krotscheck.kangaroo.database.entity.OAuthToken;
 import net.krotscheck.kangaroo.database.entity.OAuthTokenType;
 import net.krotscheck.kangaroo.servlet.oauth2.resource.TokenResponseEntity;
-import net.krotscheck.kangaroo.servlet.oauth2.util.ValidationUtil;
+import net.krotscheck.kangaroo.util.ValidationUtil;
 import org.glassfish.hk2.utilities.binding.AbstractBinder;
 import org.glassfish.jersey.process.internal.RequestScoped;
 import org.hibernate.Session;
